pyproject.toml
[tool.poetry]
name = "cp-heuristics-adapter"
version = "0.1.0"
description = ""
authors = ["suisen-cp"]
maintainers = []
license = "MIT"
readme = "README.md"
repository = "https://github.com/suisen-cp/cp-heuristics-adapter"

[tool.poetry.scripts]
cp-heuristics-adapter = "cp_heuristics_adapter.main:main"

[tool.poetry.dependencies]
python = "^3.11"
colorlog = "^6.8.2"
pyyaml = "^6.0.1"
toml = "^0.10.2"

[tool.poetry.dev-dependencies]
pytest = "^8.2.2"
mypy = "^1.10.0"
ruff = "^0.4.8"

[tool.poetry.group.dev.dependencies]
types-pyyaml = "^6.0.12.20240311"
types-toml = "^0.10.8.20240310"
pre-commit = "^3.7.1"
pytest-factoryboy = "^2.7.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
show_error_context = true
show_column_numbers = true
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
check_untyped_defs = true

[tool.ruff]
exclude = [
    ".venv",
    "venv",
    "__pycache__",
    ".git",
]

line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    "C",
    "E",
    "F",
    "W",
]
ignore = ["E501"]
fixable = ["ALL"]


[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint.mccabe]
# C901
max-complexity = 10
